solution commit
I originally was trying to do if (i.online === true) which was throwing an error stating the property 'online' was not defined for users.  This confused me until I found a partial code on the forums that pointed out that you need to store the name of the first layer of the obj (in this case it's the name of the users) and then I could call the property 'online' to evaluate.